removed quill; initial page rendering
