+++ apps/pages/package.json
@@ -10,7 +10,6 @@
     "checkFormat": "prettier src  --check --config ../.prettierrc.js"
   },
   "dependencies": {
-    "@vueup/vue-quill": "^1.0.0",
     "core-js": "3.34.0",
     "graphql-request": "5.2.0",
     "molgenis-components": "*",
